Address Canvas readback and drawImage review feedback
Read getImageData regions from the rendered framebuffer rather than the
partial CPU mirror. Preserve straight alpha, transparent out-of-bounds
pixels, resize behavior, and wide normalized region coordinates; surface
readback errors as JavaScript exceptions.

Validate drawImage arity and normalize geometry before source readback
or uploads. Coerce coordinates once before observing source dimensions,
and retain images referenced by queued draws through the NanoVG flush.

Keep PNG encoding available independently of input-image loading and
correct the imported texture-ID namespace comment.

Replace probe-canvas assertions with direct getImageData coverage and add
native no-op allocation/readback and PNG encoding regressions. Verify the
native tests with image loading enabled and disabled.

Copilot-Session: 48fa7ec8-0cf6-48a0-b836-10b7ace38c1a

+++ Polyfills/Canvas/Source/ImageData.h
@@ -9,7 +9,7 @@ namespace Babylon::Polyfills::Internal
     class ImageData final : public Napi::ObjectWrap<ImageData>
     {
     public:
-        static Napi::Value CreateInstance(Napi::Env env, Context* context, int32_t sx, int32_t sy, uint32_t width, uint32_t height);
+        static Napi::Value CreateInstance(Napi::Env env, Context* context, int64_t sx, int64_t sy, uint32_t width, uint32_t height);
 
         explicit ImageData(const Napi::CallbackInfo& info);
